Specify minimum Node version and entry point
Langchain needs polyfilling and additional dependencies for < Node v18.
This support will also be deprecated in due course.

Without specifying the correct Node version in package.json, the
deployment host will fail to utilise the host-specific settings and
overwrite them with the "missing" details in package.json, which it will
interpret as meaning "use default", thus overwriting any host-side
changes with Node v16.17.0, and an entry point of `/index.js`.